==> analysis/01-nlp-dataloading-merging.ipynb <==
# Setup

## Load in Packages

# import libraries to navigate within directories
import os
import sys
import re

# the usual suspects
import pandas as pd
import numpy as np

# Jupyter-specific imports
from IPython.display import clear_output
%matplotlib inline 

# Suppress warnings (optional)
import warnings
warnings.filterwarnings("ignore", category=DeprecationWarning)

## Access Local Git Repo


# Directory & file hierarchy
proj_dir = os.path.abspath('../')
analysis_dir = os.getcwd()
data_dir = os.path.join(proj_dir,'data')
transcript_dir = os.path.join(data_dir, 'transcripts')
results_dir = os.path.join(proj_dir,'results')
plot_dir = os.path.join(results_dir,'plots')
csv_dir = os.path.join(results_dir,'csv')

# Add helpers to python path
if os.path.join(proj_dir,'analysis','python') not in sys.path:
    sys.path.append(os.path.join(proj_dir,'analysis','python'))
    
if not os.path.exists(results_dir):
    os.makedirs(results_dir)
    
if not os.path.exists(plot_dir):
    os.makedirs(plot_dir)   
    
if not os.path.exists(csv_dir):
    os.makedirs(csv_dir)       

## Import Metadata and Transcript data from repo

### Import Metadata

# Define the path to the data directory relative to your current script
data_dir = os.path.join("..", "..", "data") # go up two directory levels

# Importing data for all channels
all_channels = pd.read_csv(os.path.join(data_dir,"master_metadata.csv"))

# Remove the 'Unnamed: 9' column if it exists
if 'Unnamed: 9' in all_channels.columns:
    all_channels = all_channels.drop(columns=['Unnamed: 9'])

# Make tier an integer
all_channels['Tier (1, 2, or 3)'] = all_channels['Tier (1, 2, or 3)'].astype(int)

# Sort to place BGC at the top
all_channels = all_channels.sort_values(
    by='channel_name',
    key=lambda x: x == 'Black Girls Code',
    ascending=[False]
)

all_channels.head(3) # let's take a peek...

### Import Transcripts

def load_txt_files_to_dataframe(root_folder):
    # Initialize an empty DataFrame to hold the content of all .txt files
    all_data = []

    # Traverse the directory tree
    for subdir, _, files in os.walk(root_folder):
        if subdir.endswith("_for_machines"):
            # Extract the channel name (so transcripts can be properly matched in cases of videos with the same name)
            channel_name = os.path.basename(subdir).replace('_transcripts_for_machines', '')

            # Process each .txt file in the folder
            for file in files:
                if file.endswith(".txt"):
                    file_path = os.path.join(subdir, file)
                    with open(file_path, 'r', encoding='utf-8') as f:
                        content = f.read()
                        # Append the content and file path to the list
                        all_data.append({'transcript_filename': file, 'content': content, 'channel_name_processed': channel_name})

    # Convert the list to a DataFrame
    df = pd.DataFrame(all_data)
    return df

transcripts = load_txt_files_to_dataframe(data_dir)

transcripts.head(3) # take a peek... yooooo, Roblox???? 

# Dataset Pre-processing: Merge Metadata and Transcript Data, then save the .csv

## Merge Metadata and Transcript data:

# Process the channel names in 'all_channels' so that they match the format of those in 'trancripts' df (to allow for merge)
all_channels['channel_name_processed'] = all_channels['channel_name'].str.lower().str.replace(':', '-') # convert to lowercase, also change ':' to '-' (due to kidicode ninja channel)

# Create a dataframe that merges metadata and transcript content
df_metadata_transcript = pd.merge(all_channels, transcripts, on=['transcript_filename','channel_name_processed'], how='left') # merge on basis of filename and channel name

# Drop processed channel name (no longer relevant)
df_metadata_transcript = df_metadata_transcript.drop(columns=['channel_name_processed'])

# View merged df
df_metadata_transcript.head(5)

## Ensure no rows have been lost or gained

Check that the lengths of the merged df, the original transcripts df, the original metadata df, are all the same as the number of transcripts in data folder.

# Count the number of files in all the folders ending in '_transcripts' 
total_files = 0

# Walk through the directory tree
for subdir, dirs, files in os.walk(data_dir):
    # Check if the folder name ends with 'for_humans'
    if subdir.endswith('for_humans'):
        # Count the number of files in this folder
        total_files += len(files)

if len(df_metadata_transcript) == len(transcripts) == len(all_channels) == total_files:
    print(f'All lengths the same! There are {len(df_metadata_transcript)} videos with transcripts.')
else:
    print('Discrepency detected. Check for merge errors.')

## Ensure that rows without transcripts are dropped:

# Drop rows that don't have transcripts
df_metadata_transcript = df_metadata_transcript.dropna(subset = 'content')
df_metadata_transcript.shape

## Save "df_metadata_transcript" as .csv

df_metadata_transcript.to_csv('metadata_transcript.csv', index=False)

## Lines of df with wrong transcript_filename entry (Nora note to self: delete later once fixed)

all_channels[all_channels['transcript_filename'] == 'loopswiththebee.txt']

df_metadata_transcript[df_metadata_transcript['transcript_filename'] == 'loopswiththebee.txt']

df_metadata_transcript[df_metadata_transcript['transcript_filename'] == 'thehourofcodeishere.txt']

df_metadata_transcript[df_metadata_transcript['transcript_filename'] == 'sideprojectsaturday.txt']

df_metadata_transcript[df_metadata_transcript['transcript_filename'] == 'howtofadebetweenscenesinunity.txt']

df_metadata_transcript[df_metadata_transcript['transcript_filename'] == '21variablesinp5jsmousexmouseyp5jstutorial.txt']
